% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_aggData.r
\name{atl_agg_data}
\alias{atl_agg_data}
\title{Aggregate data over an interval.}
\usage{
atl_agg_data(data, interval = 60)
}
\arguments{
\item{data}{Cleaned data to aggregate. Must have a numeric column named time.}

\item{interval}{The interval in seconds over which to aggregate.}
}
\value{
A dataframe aggregated taking the mean over the interval.
}
\description{
Aggregate data over an interval.
}
