% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_plot_before_after.R
\name{atl_before_after}
\alias{atl_before_after}
\title{See the effect of applying a pre-processing function.}
\usage{
atl_before_after(data, fun, x = "x", y = "y", args = list())
}
\arguments{
\item{data}{The data passed to the pre-preprocessing function.}

\item{fun}{The pre-processing function that operates on data.}

\item{x}{The X coordinate.}

\item{y}{The Y coordinate.}

\item{args}{Arguments passed as a list to the pre-processing function.}
}
\value{
Nothing. Makes a plot.
}
\description{
See the effect of applying a pre-processing function.
}
