% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_check_data.R
\name{atl_check_data}
\alias{atl_check_data}
\title{Check data has required columns.}
\usage{
atl_check_data(data, names_expected = c("x", "y", "time"))
}
\arguments{
\item{data}{The tracking data to check for required columns. Must be in the
form of a data.frame or similar, which can be handled by the function
colnames.}

\item{names_expected}{The names expected as a character vector.
By default, checks for the column names \code{x, y, time}.}
}
\value{
None. Breaks if the data does not have required columns.
}
\description{
An internal function that checks that the data.table has the required
columns. Used within most, if not all, other atlastools functions.
}
\examples{
# basic (and only) use
\dontrun{
atl_check_data(
  data = data,
  names_expected = c("x", "y", "time")
)
}

}
\author{
Pratik R. Gupte
}
