% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_compare_tracks.R
\name{atl_compare_tracks}
\alias{atl_compare_tracks}
\title{Compare reference and processed tracks.}
\usage{
atl_compare_tracks(
  processed_data,
  reference_data,
  x = "x",
  y = "y",
  time = "time"
)
}
\arguments{
\item{processed_data}{The processed movement track.}

\item{reference_data}{The reference movement track.}

\item{x}{The X coordinate.}

\item{y}{The Y coordinate.}

\item{time}{The time coordinate.}
}
\value{
The median of the ratio of the displacement between the processed
track and the reference track, and the corresponding distance between 
consecutive points in the reference track.
}
\description{
Compare reference and processed tracks.
}
