% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_filter_by_bounds.R
\name{atl_filter_bounds}
\alias{atl_filter_bounds}
\title{Remove positions within a bounding box.}
\usage{
atl_filter_bounds(
  data,
  x = "x",
  y = "y",
  x_range = c(639470, 639472),
  y_range = c(5887143, 5887145),
  remove_inside = TRUE
)
}
\arguments{
\item{data}{A dataframe or extension which contains X and Y coordinates.}

\item{x}{The X coordinate column.}

\item{y}{The Y coordinate column.}

\item{x_range}{The range of X coordinates.}

\item{y_range}{The range of Y coordinates.}

\item{remove_inside}{Whether to remove points from within the range.
Setting \code{negate = TRUE} removes positions within the bounding
box specified by the X and Y ranges.}
}
\value{
A data frame of tracking locations with attractor points removed.
}
\description{
Remove positions within a bounding box.
}
