% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_filter_covariates.R
\name{atl_filter_covariates}
\alias{atl_filter_covariates}
\title{Filter data by position covariates.}
\usage{
atl_filter_covariates(data, filters = c())
}
\arguments{
\item{data}{A dataframe or similar containing the variables to be filtered.}

\item{filters}{A character vector of filter expressions. An example might be
\code{"speed < 20"}. The filtering variables must be in the dataframe.
The function will not explicitly check whether the filtering variables are
present; this makes it flexible, allowing expressions such as
\code{"between(speed, 2, 20)"}, but also something to use at your own risk.
A missing filter variables \emph{will} result in an empty data frame.}
}
\value{
A dataframe filtered using the filters specified.
}
\description{
Filter data by position covariates.
}
