% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get_patch_data.R
\name{atl_get_patch_summary}
\alias{atl_get_patch_summary}
\title{Get residence patch data.}
\usage{
atl_get_patch_summary(
  res_patch_data,
  which_data = "summary",
  buffer_radius = 10
)
}
\arguments{
\item{res_patch_data}{A tibble with a nested list column of the raw data
underlying each patch. This column is specified in the
argument \code{dataColumn}}

\item{which_data}{Which data to return, the raw data underlying the patch,
or a spatial features object with only the patch summary.}

\item{buffer_radius}{Spatial buffer radius (in metres) around points when
requesting sf based polygons.}
}
\value{
An object of type \code{sf} or \code{tibble} depending on
which data is requested.
}
\description{
Get residence patch data.
}
