% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get_patch_data.R
\name{atl_patch_summary}
\alias{atl_patch_summary}
\title{Get residence patch data.}
\usage{
atl_patch_summary(patch_data, which_data = "summary", buffer_radius = 10)
}
\arguments{
\item{patch_data}{A data.frame with a nested list column of the raw data
underlying each patch. Since data.frames don't support nested columns,
will actually be a data.table or similar extension.}

\item{which_data}{Which data to return, the raw data underlying the patch,
or a spatial features object with only the patch summary.}

\item{buffer_radius}{Spatial buffer radius (in metres) around points when
requesting sf based polygons.}
}
\value{
An object of type \code{sf} or \code{data.table} depending on
which data is requested.
}
\description{
Get residence patch data.
}
