% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_remove_reflections.R
\name{atl_remove_reflections}
\alias{atl_remove_reflections}
\title{Remove reflected positions.}
\usage{
atl_remove_reflections(
  data,
  x = "x",
  y = "y",
  time = "time",
  point_angle_cutoff = 45,
  reflection_speed_cutoff = 20,
  est_ref_len = 1000
)
}
\arguments{
\item{data}{A dataframe or similar which has previously been cleaned.}

\item{x}{The name of the X coordinate column.}

\item{y}{The name of the Y coordinate column.}

\item{time}{The name of the timestamp column.}

\item{point_angle_cutoff}{The turning angle (in degrees) above which
high instantaneous speeds are considered an anomaly rather than fast transit.}

\item{reflection_speed_cutoff}{The speed (in m/s) above which an anomaly is
detected when combined with a high turning angle.}

\item{est_ref_len}{How many positions are expected to be in a reflection.}
}
\value{
A dataframe with reflections removed.
}
\description{
Remove reflections, or prolonged spikes from a movement track by identifying the bounds and removing positions between them.
The important function arguments here are \code{point_angle_cutoff} ($A$), \code{reflection_speed_cutoff} ($S$), and \code{est_ref_len}, the maximum number of positions after the anchor point that are candidates for the end of the prolonged spike. 
If the prolonged spike ends after less than N positions, the true end point is used as the outer bound of the spike.
However, the algorithm behind this function fails when the prolonged spike ends after more than N positions. Users are advised to use a liberally large value of N in the \code{est_ref_len} argument; 1,000 may be appropriate.
}
\examples{
\dontrun{
filtered_data <- atl_remove_reflections(data = track_data,
                      x = "x", y = "y", time = "time",
                      point_angle_cutoff = A,
                      reflection_speed_cutoff = S,
                      est_ref_len = N)
}
}
\author{
Pratik R. Gupte
}
