% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_filter_by_bbox.r
\name{atl_rm_attractor}
\alias{atl_rm_attractor}
\title{Remove positions within a bounding box.}
\usage{
atl_rm_attractor(
  df,
  atp_xmin = 639470,
  atp_xmax = 639472,
  atp_ymin = 5887143,
  atp_ymax = 5887145
)
}
\arguments{
\item{df}{A dataframe or extension which contains capitalised X and Y coordinates.}

\item{atp_xmin}{The min X coordinates of attractor locations.}

\item{atp_xmax}{The max X coordinates of attractor locations.}

\item{atp_ymin}{The min Y coordinates of attractor locations.}

\item{atp_ymax}{The max Y coordinates of attractor locations.}
}
\value{
A data frame of tracking locations with attractor points removed.
}
\description{
Remove positions within a bounding box.
}
