% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_simple_distance.R
\name{atl_simple_dist}
\alias{atl_simple_dist}
\title{Calculate euclidean distances between successive points.}
\usage{
atl_simple_dist(data, x = "x", y = "y")
}
\arguments{
\item{data}{A dataframe object of or extending the class data.frame, which must contain at least two coordinate columns for the X and Y coordinates.}

\item{x}{A column name in a data.frame object that contains the numeric X or longitude coordinate for position data.}

\item{y}{A column name in a data.frame object that contains the numeric Y or latitude coordinate for position data.}
}
\value{
Returns a vector of distances between consecutive points.
}
\description{
Calculate euclidean distances between successive points.
}
