% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_getData.r
\name{funcGetData}
\alias{funcGetData}
\title{getData}
\usage{
funcGetData(
  tag,
  tracking_time_start,
  tracking_time_end,
  tag_prefix = "31001000",
  database = "db",
  host = "abtdb1.nioz.nl",
  username = "someuser",
  password = "somepassword"
)
}
\arguments{
\item{tag}{A three digit numeric representing the ToA tag.}

\item{tracking_time_start}{Character representation of time (Central European Time, +1 UTC) from which start point data should be retrieved.}

\item{tracking_time_end}{Character time representing the end point corresponding to the above start point.}

\item{tag_prefix}{A numeric specifying the tag prefix. Defaults to \code{31001000}}

\item{database}{The database name on the host server. Defaults to \code{db} for unknown reasons.}

\item{host}{The server address on which the data are stored. Defaults to \code{abtdb1.nioz.nl}.}

\item{username}{Username to access the data.}

\item{password}{Password to access the data. Contact \code{allert.bijleveld@nioz.nl} to get access.}
}
\value{
A dataframe of agent positions, filtered between the required tracking times.
}
\description{
getData
}
