% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_getPatchData.r
\name{funcGetPatchData}
\alias{funcGetPatchData}
\title{getPatchData}
\usage{
funcGetPatchData(resPatchData, dataColumn = "data", whichData = "spatial")
}
\arguments{
\item{resPatchData}{A tibble with a nested list column of the raw data underlying each patch. This column is specified in the argument \code{dataColumn}}

\item{dataColumn}{Which nested list column contains the data.}

\item{whichData}{Which data to return, the raw data underlying the patch, or a spatial features object with only the patch summary.}
}
\value{
An object of type \code{sf} or \code{tibble} depending on which data is requested.
}
\description{
getPatchData
}
