% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_getResPatch.r
\name{funcGetResPatch}
\alias{funcGetResPatch}
\title{getResPatch}
\usage{
funcGetResPatch(somedata, bufferSize = 10, spatIndepLim = 50,
  tempIndepLim = 3600, makeSf = FALSE)
}
\arguments{
\item{somedata}{A dataframe of values of any class that is or extends data.frame. The dataframe must contain at least two spatial coordinates, \code{x} and \code{y}, and a temporal coordinate, \code{time}. The names of columns specifying these can be passed as arguments below.}

\item{bufferSize}{A numeric value specifying the radius of the buffer to be considered around each coordinate point. May be thought of as the distance that an individual can access, assess, or otherwise cover when at a discrete point in space.}

\item{spatIndepLim}{A numeric value of time in seconds of the time difference between two patches for them to be considered independent.}

\item{tempIndepLim}{A numeric value of distance in metres of the spatial distance between two patches for them to the considered independent.}

\item{makeSf}{Whether to return an sf object rather than simply a dataframe, the geometry being the patch outlines.}
}
\value{
A data.frame extension object. This dataframe has the added column \code{resPatch} based on cumulative patch summing. Depending on whether \code{inferPatches = TRUE}, the dataframe has additional inferred points. An additional column is created in each case, indicating whether the data are empirical fixes ('real') or 'inferred'.
}
\description{
getResPatch
}
