% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_residencePatch.r
\name{funcGetResPatches}
\alias{funcGetResPatches}
\title{getResPatches}
\usage{
funcGetResPatches(df, x = "x", y = "y", time = "time",
  buffsize = 10, returnSf = FALSE)
}
\arguments{
\item{df}{A dataframe of values of any class that is or extends data.frame. The dataframe must contain at least two spatial coordinates, `x` and `y`, and a temporal coordinate, `time`. The names of columns specifying these can be passed as arguments below.}

\item{x}{A column name (string) containing X (or longitude) coordinates.}

\item{y}{A column name (string) containing Y (or latitude) coordinates.}

\item{time}{A column name (string) containing the timestamp associated with each spatial coordinate pair.}

\item{buffsize}{A numeric value specifying the radius of the buffer to be drawn around each coordinate point. May be thought of as the distance that an individual can access, assess, or otherwise cover when at a discrete point in space.}

\item{returnSf}{A logical value (`TRUE` or `FALSE`) of whether the constructed patches should be returned. When true, the `sf` object is returned bound to the patch summary dataframe in a list of length two.}
}
\value{
Depending on whether the constructed poylgons are requested, a list object containing as its first element a dataframe of patch summaries, and as its second element an `sf` `MULTIPOLYGON` or `POLYGON` (as approporiate) object. The list is returned when `returnSf = TRUE`, and only the dataframe of patch summaries when `returnSf` is `FALSE`.
}
\description{
getResPatches
}
\author{
Pratik R Gupte, \email{p.r.gupte@rug.nl}
}
