% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_inferResidence.r
\name{funcInferResidence}
\alias{funcInferResidence}
\title{segPath2}
\usage{
funcInferResidence(revdata, htdata, resTimeLimit = 2, travelSeg = 5,
  infPatchTimeDiff = 1800, infPatchSpatDiff = 100)
}
\arguments{
\item{revdata}{A string filepath to data in csv format. The data must be the output of recurse analysis, or must include, in addition to X, Y and time columns, a residence time column named resTime, id, and tidalcycle.}

\item{htdata}{A string filepath to data in csv format. The data must be the output of tidal cycle finding analysis, or must include, in addition to X, Y and time columns, a tidaltime column named tidaltime; also id, and tidalcycle for merging.}

\item{resTimeLimit}{A numeric giving the time limit in minutes against which residence time is compared.}

\item{travelSeg}{A numeric value of the number of fixes, or rows, over which a smoother function is applied.}

\item{infPatchTimeDiff}{A numeric duration in seconds, of the minimum time difference between two points, above which, it is assumed worthwhile to examine whether there is a missing residence patch to be inferred.}

\item{infPatchSpatDiff}{A numeric distance in metres, of the maximum spatial distance between two points, below which it may be assumed few extreme movements took place between them.}
}
\value{
A data.frame extension object. This dataframe has additional inferred points, indicated by the additional column for empirical fixes ('real') or 'inferred'.
}
\description{
segPath2
}
