% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_trajBwPatch.r
\name{funcPatchTraj}
\alias{funcPatchTraj}
\title{patchTraj}
\usage{
funcPatchTraj(df, x1 = "X_end", x2 = "X_start", y1 = "Y_end",
  y2 = "Y_start")
}
\arguments{
\item{df}{A dataframe with at least 2 pairs of coordinate columns, such as X_end Y_end, X_start, Y_start.}

\item{x1}{Coordinate column for the beginning X coordinate of the linestring; this should be the last X coordinate of each patch if inter-patch linestrngs are to be constructed. It is possible to use the X_start coordinate, but this will lead to issues and nonsensical output at this point.}

\item{x2}{Coordinate column for the end X coordinate of the linestring; this should be the first X coordinate of each patch if inter-patch linestrngs are to be constructed.}

\item{y1}{Coordinate column for the beginning Y coordinate of the linestring; this should be the last Y coordinate of each patch if inter-patch linestrngs are to be constructed. It is possible to use the X_start coordinate, but this will lead to issues and nonsensical output at this point.}

\item{y2}{Coordinate column for the end Y coordinate of the linestring; this should be the first Y coordinate of each patch if inter-patch linestrngs are to be constructed.}
}
\value{
An `sf` `MULTILINESTRING` object representing the linear path between N patches, starting at the end point of the first patch, and ending at the first point of the Nth patch.
}
\description{
patchTraj
}
