% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_segmentPath.r
\name{funcSegPath}
\alias{funcSegPath}
\title{segPath}
\usage{
funcSegPath(revdata, resTimeLimit = 2, travelSeg = 5,
  inferPatches = TRUE)
}
\arguments{
\item{revdata}{A string filepath to data in csv format. The data must be the output of recurse analysis, or must include, in addition to X, Y and time columns, a residence time column named resTime, id, and tidalcycle.}

\item{resTimeLimit}{A numeric giving the time limit in minutes against which residence time is compared.}

\item{travelSeg}{A numeric value of the number of fixes, or rows, over which a smoother function is applied.}

\item{inferPatches}{A logical indicating whether residence patches should be inferred from temporal gaps in the data.}
}
\value{
A data.frame extension object. This dataframe has the added column `resPatch` based on cumulative patch summing. Depending on whether `inferPatches` is set `TRUE`, the dataframe has additional inferred points. An additional column is created in each case, indicating whether the data are empirical fixes ('real') or 'inferred'.
}
\description{
segPath
}
