% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_aggData.r
\name{wat_agg_data}
\alias{wat_agg_data}
\title{A function to aggregate WATLAS data over an interval.}
\usage{
wat_agg_data(data, interval = 60)
}
\arguments{
\item{data}{Cleaned data - the output of the cleanData function - to aggregate.
Must have a numeric column named time.}

\item{interval}{The interval in seconds over which to aggregate.}
}
\value{
A dataframe aggregated taking the mean over the interval.
}
\description{
A function to aggregate WATLAS data over an interval.
}
