% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_classifyPath.r
\name{wat_classify_points}
\alias{wat_classify_points}
\title{Classify points as residence points based on residence time.}
\usage{
wat_classify_points(somedata, resTimeLimit = 2)
}
\arguments{
\item{somedata}{A data frame which must have a column for the residence time at each point.}

\item{resTimeLimit}{A numeric giving the time limit in minutes against which residence time is compared.}
}
\value{
A data.frame extension object, which retains only points classified as residence points if residence time is below \code{resTimeLimit} over \code{travelSeg} points.
}
\description{
Classify points as residence points based on residence time.
}
