% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_repair_ht_patches.r
\name{wat_repair_ht_patches}
\alias{wat_repair_ht_patches}
\title{repair high tide patches}
\usage{
wat_repair_ht_patches(patch_data_list, spatIndepLim = 100,
  tempIndepLim = 30, bufferSize = 10)
}
\arguments{
\item{patch_data_list}{A list of data.tables, each the output of make_res_patch. Must have an sfc geometry column.}

\item{spatIndepLim}{The spatial independence limit.}

\item{tempIndepLim}{The temporal independence limit.}

\item{bufferSize}{The buffer size for spatial polygons.}
}
\value{
A datatable with repaired high tide patches.
}
\description{
repair high tide patches
}
