% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_rmAttractor.r
\name{wat_rmAttractor}
\alias{wat_rmAttractor}
\title{A function to remove so-called attractor points which have an unknown source. To be run before cleaning the data.}
\usage{
wat_rmAttractor(
  df,
  atp_xmin = 639470,
  atp_xmax = 639472,
  atp_ymin = 5887143,
  atp_ymax = 5887145
)
}
\arguments{
\item{df}{A dataframe or extension which contains capitalised X and Y coordinates.}

\item{atp_xmin}{The min X coordinates of attractor locations.}

\item{atp_xmax}{The max X coordinates of attractor locations.}

\item{atp_ymin}{The min Y coordinates of attractor locations.}

\item{atp_ymax}{The max Y coordinates of attractor locations.}
}
\value{
A data frame of tracking locations with attractor points removed.
}
\description{
A function to remove so-called attractor points which have an unknown source. To be run before cleaning the data.
}
