% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_check_data.R
\name{atl_check_data}
\alias{atl_check_data}
\title{Check data has required columns.}
\usage{
atl_check_data(data, names_expected = c("x", "y", "time"))
}
\arguments{
\item{data}{The tracking data to check for required columns.}

\item{names_expected}{The names expected.}
}
\value{
None. Breaks if the data does not have required columns.
}
\description{
Check data has required columns.
}
