% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get_speed.R
\name{atl_get_speed}
\alias{atl_get_speed}
\title{Calculate instantaenous speed.}
\usage{
atl_get_speed(data, x = "x", y = "y", time = "time", type = c("in"))
}
\arguments{
\item{data}{A dataframe or similar which must have the columns
specified by \code{x}, \code{y}, and \code{time}.}

\item{x}{The x coordinate.}

\item{y}{The y coordinate.}

\item{time}{The timestamp in seconds since the UNIX epoch.}

\item{type}{The type of speed (incoming or outgoing) to return.
Incoming speeds are specified by \code{type = "in"}, and outgoing speeds
by \code{type = "out"}.}
}
\value{
A vector of numerics representing speed.
The first position is assigned a speed of NA.
}
\description{
Calculate instantaenous speed.
}
