% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_remove_reflections.R
\name{atl_remove_reflections}
\alias{atl_remove_reflections}
\title{Remove reflected positions.}
\usage{
atl_remove_reflections(
  data,
  x = "x",
  y = "y",
  time = "time",
  point_angle_cutoff = 45,
  reflection_speed_cutoff = 40,
  est_ref_len = 1000
)
}
\arguments{
\item{data}{A dataframe or similar which has previously been cleaned.}

\item{x}{The name of the X coordinate column.}

\item{y}{The name of the Y coordinate column.}

\item{time}{The name of the timestamp column.}

\item{point_angle_cutoff}{The turning angle (in degrees) above which
high instantaneous speeds are considered an anomaly rather than fast transit.}

\item{reflection_speed_cutoff}{The speed (in m/s) above which an anomaly is
detected when combined with a high turning angle.}
}
\value{
A dataframe with reflections removed.
}
\description{
Remove reflected positions.
}
\author{
Pratik R. Gupte
}
