% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_get_angle.R
\name{atl_turning_angle}
\alias{atl_turning_angle}
\title{Get the turning angle between points.}
\usage{
atl_turning_angle(data, x = "x", y = "y", time = "time")
}
\arguments{
\item{data}{A dataframe or similar which must have the columns
specified by \code{x}, \code{y}, and \code{time}.}

\item{x}{The x coordinate.}

\item{y}{The y coordinate.}

\item{time}{The timestamp in seconds since the UNIX epoch.}
}
\value{
A vector of turning angles in degrees.
Negative degrees indicate 'left' turns. There are two fewer
angles than the number of rows in the dataframe.
}
\description{
Get the turning angle between points.
}
