% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_classifyPath.r
\name{funcClassifyPath}
\alias{funcClassifyPath}
\title{classifyPath}
\usage{
funcClassifyPath(somedata, restimeCol = "resTime", resTimeLimit = 2,
  smoother = 5)
}
\arguments{
\item{somedata}{A data frame which must have a column for the residence time at each point.}

\item{restimeCol}{The residence time column name.}

\item{resTimeLimit}{A numeric giving the time limit in minutes against which residence time is compared.}

\item{smoother}{A numeric value of the number of fixes, or rows, over which a smoother function is applied.}
}
\value{
A data.frame extension object, which retains only points classified as residence points if residence time is below \code{resTimeLimit} over \code{travelSeg} points.
}
\description{
classifyPath
}
