% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_inferResidence.r
\name{funcInferResidence}
\alias{funcInferResidence}
\title{segPath2}
\usage{
funcInferResidence(
  df,
  infResTime = 2,
  infPatchTimeDiff = 30,
  infPatchSpatDiff = 100
)
}
\arguments{
\item{df}{A dataframe of recurse analysis, or must include, in addition to x, y and time columns, a residence time column named resTime, TAG, and tidalcycle, a tidaltime column named tidaltime; also id, and tidalcycle for merging.}

\item{infResTime}{A numeric giving the time limit in minutes against which residence time is compared.}

\item{infPatchTimeDiff}{A numeric duration in minutes, of the minimum time difference between two points, above which, it is assumed worthwhile to examine whether there is a missing residence patch to be inferred.}

\item{infPatchSpatDiff}{A numeric distance in metres, of the maximum spatial distance between two points, below which it may be assumed few extreme movements took place between them.}
}
\value{
A data.frame extension object. This dataframe has additional inferred points, indicated by the additional column for empirical fixes ('real') or 'inferred'.
}
\description{
segPath2
}
