% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_runApp.r
\name{wat_run_patch_vis_app}
\alias{wat_run_patch_vis_app}
\title{Run Shiny app to run examples of the patch functions.}
\usage{
wat_run_patch_vis_app()
}
\value{
Opens a shiny app that allows tweaking of patch construction parameters.
}
\description{
Run Shiny app to run examples of the patch functions.
}
